+++ Core/Component/Component.h
@@ -48,8 +48,8 @@ public:
 	void ObjectInitialized();
 
 	virtual void PhysxUpdate() {}
-	virtual void Update( float ElapsedTime ) {}
-	virtual void FixedUpdate( float ElapsedTime ) {}
+	virtual void Update( double ElapsedTime ) {}
+	virtual void FixedUpdate( double ElapsedTime ) {}
 	virtual void Render(CContextManager&  _context) {}
 	virtual void RenderDebug(CContextManager&  _context) {}
